+++ Obsidian/WorldData/Generators/Overworld/Decorators/SnowyTundraDecorator.cs
@@ -13,8 +13,8 @@ namespace Obsidian.WorldData.Generators.Overworld.Decorators
 
         public override void Decorate(Chunk chunk, Position pos, OverworldNoise noise)
         {
-            int worldX = (chunk.X << 4) + (int) pos.X;
-            int worldZ = (chunk.Z << 4) + (int) pos.Z;
+            int worldX = (chunk.X << 4) + pos.X;
+            int worldZ = (chunk.Z << 4) + pos.Z;
             
             var sand = Registry.GetBlock(Materials.SnowBlock);
             var sandstone = Registry.GetBlock(Materials.PackedIce);
